Server: Seperated all games related endpoints to a seperate router.
